requirements.txt
# API + workers
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
celery[redis]>=5.4.0
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.7.0
python-multipart>=0.0.9
httpx>=0.27.0
boto3>=1.34.0
pyyaml>=6.0.0

# ML core
torch>=2.3.0
torchvision>=0.18.0
diffusers>=0.29.0
transformers>=4.42.0
huggingface-hub>=0.23.0
accelerate>=0.31.0
safetensors>=0.4.0

# Image
Pillow>=10.3.0

# Video
imageio[ffmpeg]>=2.34.0
numpy>=1.26.0

# Audio
soundfile>=0.12.0
demucs>=4.0.1
audiocraft>=1.3.0
kokoro>=0.9.0
# f5-tts  # install separately: pip install git+https://github.com/SWivid/F5-TTS

# Speech recognition (for dubbing)
# whisperx  # install separately: pip install git+https://github.com/m-bain/whisperX
